Dodana możliwość chodu zamiast biegu (Lshift)

+++ postac.h
@@ -66,7 +66,7 @@ public:
 protected:
     float x_speed; //do ustawienia
     float y_speed;
-    int v_ratio; //przelicznik predkosci
+    float v_ratio; //przelicznik predkosci
     bool is_moving=false;
 };
 #endif // POSTAC_H
